+++ friends/src/app/screens/MainScreen/ManageFriendsDrawerContent/index.tsx
@@ -1,21 +1,18 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { FlatList, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
-import { useServiceContext } from '../../../../main';
-import { FriendId } from '../../../../entities/friend';
-import { useFriends } from '../../../queries/useFriends';
+import { StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import { useMainScreenContext } from '..';
+import { useState } from 'react';
+import { ManageFriendsDrawerContentActiveList } from './ManageFriendsDrawerContentActiveList';
+import { ManageFriendsDrawerContentRequestedList } from './ManageFriendsDrawerContentRequestedList';
+
+type Tab = 'ACTIVE' | 'REQUESTED';
 
 type Props = {
   onClose: () => void;
 };
 
 export const ManageFriendsDrawerContent = ({ onClose }: Props) => {
+  const [tab, setTab] = useState<Tab>('ACTIVE');
   const { onSelectFriend } = useMainScreenContext();
-  const { data: requestedFriends } = useFriends({ state: 'REQUESTED' });
-  const { data: activeFriends } = useFriends({ state: 'ACTIVE' });
-  const { mutate: acceptFriend } = useAcceptFriend();
-  const { mutate: declineFriend } = useDeclineFriend();
-  const { mutate: deleteFriend } = useDeleteFriend();
 
   return (
     <View style={styles.container}>
@@ -25,46 +22,26 @@ export const ManageFriendsDrawerContent = ({ onClose }: Props) => {
           <Text>X</Text>
         </TouchableOpacity>
       </View>
-      <Text>요청받은거</Text>
-      <FlatList
-        data={requestedFriends}
-        renderItem={({ item }) => (
-          <View>
-            <Text>
-              {item.nickname}#{item.tag}
-            </Text>
-            <TouchableOpacity onPress={() => acceptFriend(item.friendId)}>
-              <Text>승인</Text>
+      <View>
+        <View>
+          {tabs.map(({ label, value }) => (
+            <TouchableOpacity key={value} onPress={() => setTab(value)}>
+              <View>
+                <Text>{label}</Text>
+              </View>
             </TouchableOpacity>
-            <TouchableOpacity onPress={() => declineFriend(item.friendId)}>
-              <Text>거절</Text>
-            </TouchableOpacity>
-          </View>
-        )}
-      />
-      <Text>친구 목록</Text>
-      <FlatList
-        data={activeFriends}
-        renderItem={({ item }) => (
-          <View>
-            <TouchableOpacity
-              onPress={() => {
-                onSelectFriend(item.friendId);
-                onClose();
-              }}
-            >
-              <Text>
-                {item.nickname}#{item.tag}
-              </Text>
-            </TouchableOpacity>
-            <TouchableOpacity
-              onPress={() => deleteFriend(item.friendId, { onSuccess: () => onSelectFriend(undefined) })}
-            >
-              <Text>삭제</Text>
-            </TouchableOpacity>
-          </View>
-        )}
-      />
+          ))}
+        </View>
+        <View>
+          <Text>친구 추가하기 +</Text>
+        </View>
+        {
+          {
+            ACTIVE: <ManageFriendsDrawerContentActiveList onClickFriend={onSelectFriend} />,
+            REQUESTED: <ManageFriendsDrawerContentRequestedList />,
+          }[tab]
+        }
+      </View>
     </View>
   );
 };
@@ -78,26 +55,7 @@ const styles = StyleSheet.create({
   },
 });
 
-const useAcceptFriend = () => {
-  const queryClient = useQueryClient();
-  const { friendService } = useServiceContext();
-  return useMutation((friendId: FriendId) => friendService.acceptFriend({ friendId }), {
-    onSuccess: () => queryClient.invalidateQueries(),
-  });
-};
-
-const useDeclineFriend = () => {
-  const queryClient = useQueryClient();
-  const { friendService } = useServiceContext();
-  return useMutation((friendId: FriendId) => friendService.declineFriend({ friendId }), {
-    onSuccess: () => queryClient.invalidateQueries(),
-  });
-};
-
-const useDeleteFriend = () => {
-  const queryClient = useQueryClient();
-  const { friendService } = useServiceContext();
-  return useMutation((friendId: FriendId) => friendService.deleteFriend({ friendId }), {
-    onSuccess: () => queryClient.invalidateQueries(),
-  });
-};
+const tabs: { label: string; value: Tab }[] = [
+  { label: '친구 목록', value: 'ACTIVE' },
+  { label: '친구 요청', value: 'REQUESTED' },
+];
